Cleaner YAML with ansible blocks
The ansible block feature is now used. First of, it enables fewer
lines, by moving tags, become and when to a single place.

Further, it now uses when: variable is defined type conditions instead
of empty lists by default. As a result skipped tasks will now be
properly reported, yet the chance of accidental undefined errors should
be small.