Deprecate ability to define plain text password
It is no longer possible to define plain text passwords. Instead only
the recommended way is supported. Now the user must first generate a
hashed and salted password and use that directly.

There is one minor security downside to this though. It completly
disables the ability to have a unique hash for each host. Although
this was only possible by making the salt pre-defined, which itself
is less secure, even if that to were host specific.

In any case the risk of accidentaly leaking plain text passwords is
too great.

The best way to manage passwords is probably to set an initial password
using ansible and force a user to change it upon login.

Tests and documentation is updated accordingly.